Security fix

+++ src/Controller/ProductController.php
@@ -40,9 +40,12 @@ class ProductController extends AbstractController
 {
     /**
      * @Route("/", name="product_index", methods={"GET"})
+     * @param ProductRepository $productRepository
+     * @return Response
      */
     public function index(ProductRepository $productRepository): Response
     {
+        $this->denyAccessUnlessGranted("ROLE_USER");
         return $this->render('product/index.html.twig', [
             'products' => $productRepository->findAll(),
         ]);
@@ -50,9 +53,12 @@ class ProductController extends AbstractController
 
     /**
      * @Route("/new", name="product_new", methods={"GET","POST"})
+     * @param Request $request
+     * @return Response
      */
     public function new(Request $request): Response
     {
+        $this->denyAccessUnlessGranted("ROLE_USER");
         $product = new Product();
         $form = $this->createForm(Product1Type::class, $product);
         $form->handleRequest($request);
@@ -73,9 +79,12 @@ class ProductController extends AbstractController
 
     /**
      * @Route("/{id}", name="product_show", methods={"GET"})
+     * @param Product $product
+     * @return Response
      */
     public function show(Product $product): Response
     {
+        $this->denyAccessUnlessGranted("ROLE_USER");
         return $this->render('product/show.html.twig', [
             'product' => $product,
         ]);
@@ -83,9 +92,13 @@ class ProductController extends AbstractController
 
     /**
      * @Route("/{id}/edit", name="product_edit", methods={"GET","POST"})
+     * @param Request $request
+     * @param Product $product
+     * @return Response
      */
     public function edit(Request $request, Product $product): Response
     {
+        $this->denyAccessUnlessGranted("ROLE_USER");
         $form = $this->createForm(Product1Type::class, $product);
         $form->handleRequest($request);
 
@@ -103,9 +116,13 @@ class ProductController extends AbstractController
 
     /**
      * @Route("/{id}", name="product_delete", methods={"DELETE"})
+     * @param Request $request
+     * @param Product $product
+     * @return Response
      */
     public function delete(Request $request, Product $product): Response
     {
+        $this->denyAccessUnlessGranted("ROLE_USER");
         if ($this->isCsrfTokenValid('delete'.$product->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($product);
